+++ analysis/analyze_and_save_ia.R
@@ -5,12 +5,13 @@ suppressPackageStartupMessages(library("stringr"))
 ================================================================================
 Parse feather formatted data, analyze the index of association and save the result
 
-Usage: analyze_and_save_ia.R [-dvkc -s SEED [-n NSAMPLE...] -p PERMUTATIONS -l LOCUS -o PATH] [FILE...]
+Usage: analyze_and_save_ia.R [-dvkcm -s SEED [-n NSAMPLE...] -p PERMUTATIONS -l LOCUS -o PATH] [FILE...]
 
 Options:
  -h,--help                                    show this message and exit
  -v,--verbose                                 record progress
  -d,--debug                                   record EVERYTHING
+ -m,--mutant                                  flag for indicating single mutant locus. If this is flagged, two analyses are run.
  -k,--keep                                    keep the data in the resulting data frame
  -c,--clonecorrect                            add values from clone-correction
  -s SEED,--seed=SEED                          random seed [default: 20160909]
@@ -70,39 +71,67 @@ for (f in opt$FILE){
   set.seed(opt$seed) # same seed for each file
   indat <- feather2genind(f, locus_prefix = opt$locus, sample = totsamp, verbose = opt$verbose) %>%
     setPop(~pop)
-
   inpops  <- nPop(indat)
   allpops <- inpops * length(opt$nsample)
   subpops <- rep(opt$nsample, inpops) %>%
     rep(., .) %>%
     paste0("sam_", .)
-
   if (opt$verbose) message(paste("Analyzing", allpops, "populations ... "))
   res <- indat %>%
     addStrata(data.frame(sample_size = subpops)) %>%
     setPop(~pop/sample_size) %>%
     seppop() %>%
-    map(tidy_ia, 
-        sample   = opt$permutations, 
+    map(tidy_ia,
+        sample   = opt$permutations,
         keepdata = opt$keep,
-        cc       = opt$clonecorrect, 
-        strata   = ~pop/sample_size, 
-        verbose  = opt$verbose, 
-        hist     = FALSE, 
+        cc       = opt$clonecorrect,
+        strata   = ~pop/sample_size,
+        verbose  = opt$verbose,
+        hist     = FALSE,
         quiet    = !opt$debug) %>%
     bind_rows()
-  outf     <- make.names(f)
-  outfDATA <- paste0(outf, ".DATA")
-  resDATA  <- res %>% select(dataset, pop)
-  res      <- res %>% select(-dataset)
+  if (opt$mutant){
+  # At this point, if the first locus is a locus with a higher mutation rate
+  # than the other loci, It is removed, MLGS recalculated, and everything is
+  # processed again.
+    if (opt$verbose) message(paste("Analyzing populations without first locus ... "))
+    mres <- indat[loc = -1, mlg.reset = TRUE] %>%
+      addStrata(data.frame(sample_size = subpops)) %>%
+      setPop(~pop/sample_size) %>%
+      seppop() %>%
+      map(tidy_ia,
+          sample   = opt$permutations,
+          keepdata = FALSE,
+          cc       = opt$clonecorrect,
+          strata   = ~pop/sample_size,
+          verbose  = opt$verbose,
+          hist     = FALSE,
+          quiet    = !opt$debug) %>%
+      bind_rows()
+  }
+  # Saving everything.
+  outf <- make.names(f)
+  if (opt$keep){
+    outfDATA <- paste0(outf, ".DATA")
+    resDATA  <- res %>% dplyr::select(matches("dataset"), matches("pop"))    
+    assign(x = outfDATA, resDATA)
+    outf_data_location <- paste0(opt$output, "/", outfDATA, ".rda")
+    if (opt$verbose) message(paste("saving data to", outf_data_location))
+    save(list = outfDATA, file = outf_data_location)
+  }
+  res  <- res %>% dplyr::select(-matches("dataset"))
   assign(x = outf, res)
-  assign(x = outfDATA, resDATA)
-  outf_location      <- paste0(opt$output, "/", outf, ".rda")
-  outf_data_location <- paste0(opt$output, "/", outfDATA, ".rda")
+  outf_location <- paste0(opt$output, "/", outf, ".rda")
   if (opt$verbose) message(paste("saving results to", outf_location))
   save(list = outf, file = outf_location)
-  if (opt$verbose) message(paste("saving data to", outf_data_location))
-  save(list = outfDATA, file = outf_data_location)
+
+  if (opt$mutant){
+    moutf <- paste0(outf, ".mutant")
+    assign(x = moutf, mres)
+    moutf_location <- paste0(opt$output, "/", moutf, ".rda")
+    if (opt$verbose) message(paste("saving mutant results to", moutf_location))
+    save(list = moutf, file = moutf_location)
+  }
 }
 if (opt$verbose){
   options(width = 100)