+++ analysis/analyze_locus_contribution.R
@@ -9,12 +9,13 @@ round-robin multilocus genotypes, diversity statistics based on the number of
 multilocus genotypes contributed by each locus, and probability of genotypes for
 each sample.
 
-Usage: analyze_locus_contribution.R [-dv -s SEED -p PERMUTATIONS -o PATH] [FILE...]
+Usage: analyze_locus_contribution.R [-dvm -s SEED -p PERMUTATIONS -o PATH] [FILE...]
 
 Options:
  -h,--help                                    show this message and exit
  -v,--verbose                                 record progress
  -d,--debug                                   record EVERYTHING
+ -m,--mutant                                  flag for indicating single mutant locus. If this is flagged, two analyses are run.
  -s SEED,--seed=SEED                          random seed [default: 20160909]
  -p PERMUTATIONS,--permutations=PERMUTATIONS  number of permutations for the index of association [default: 99]
  -o PATH,--output=PATH                        default path to place Rdata files [default: ~/rda_files]
@@ -71,6 +72,11 @@ uSimp <- function(x){
   return((N/(N-1))*lambda)
 }
 
+# wrapper to exclude first locus before analyzing diversity.
+tidy_no_first_locus <- function(x, ...){
+  tidy_locus_contribution(x[loc = -1, mlg.reset = TRUE], ...)
+}
+
 # For each file we will
 # 1. Load the file into R (since it will be an Rda file)
 # 2. Select the dataset column
@@ -89,11 +95,30 @@ for (f in opt$FILE){
           n = opt$permutations,
           verbose = opt$verbose) %>%
     bind_rows()
+  if (opt$mutant){
+    if (opt$verbose) message(paste("Analyzing populations without first locus ... "))
+    set.seed(opt$seed)
+    mres <- indat %>%
+      get() %>%
+      select(dataset) %>%
+      apply(1, listfun, tidy_no_first_locus,
+            uSimp = uSimp,
+            n = opt$permutations,
+            verbose = opt$verbose) %>%
+      bind_rows()
+  }
   outf <- gsub(".DATA", ".contrib", basename(f))
   assign(x = outf, res)
   outf_location <- file.path(opt$output, outf)
   if (opt$verbose) message(paste("saving results to", outf_location))
   save(list = outf, file = outf_location)
+  if (opt$mutant){
+    moutf <- gsub(".DATA", ".contrib.mutant", basename(f))
+    assign(x = moutf, mres)
+    moutf_location <- file.path(opt$output, moutf)
+    if (opt$verbose) message(paste("saving results to", moutf_location))
+    save(list = moutf, file = moutf_location)
+  }
 }
 if (opt$verbose){
   options(width = 100)