+++ analysis/analyze_locus_table.R
@@ -5,12 +5,13 @@ suppressPackageStartupMessages(library("docopt"))
 Parse RDA files containing genclone objects, analyze locus diversity statistics, 
 and save the resulting data frames.
 
-Usage: analyze_and_save_ia.R [-dv -o PATH] [FILE...]
+Usage: analyze_and_save_ia.R [-dvm -o PATH] [FILE...]
 
 Options:
  -h,--help              show this message and exit
  -v,--verbose           record progress
  -d,--debug             record EVERYTHING
+ -m,--mutant            flag for indicating single mutant locus. If this is flagged, two analyses are run.
  -o PATH,--output=PATH  default path to place Rdata files [default: ~/rda_files]
 
 ================================================================================
@@ -43,6 +44,11 @@ if (opt$verbose){
 suppressPackageStartupMessages(library("zksimanalysis"))
 
 
+# wrapper to exclude first locus before analyzing diversity.
+tidy_no_first_locus <- function(x, ...){
+  tidy_locus_table(x[loc = -1, mlg.reset = TRUE], ...)
+}
+
 # Processing data ---------------------------------------------------------
 
 # For each file we will
@@ -60,11 +66,27 @@ for (f in opt$FILE){
     select(dataset) %>% 
     apply(1, listfun, tidy_locus_table,  verbose = opt$debug) %>% 
     bind_rows()
+  if (opt$mutant){
+    if (opt$verbose) message(paste("Analyzing populations without first locus ... "))
+    set.seed(opt$seed)
+    mres <- indat %>% 
+      get() %>%
+      select(dataset) %>% 
+      apply(1, listfun, tidy_no_first_locus,  verbose = opt$debug) %>% 
+      bind_rows()
+  }
   outf <- gsub(".DATA", ".locustable", basename(f))
   assign(x = outf, res)
   outf_location <- file.path(opt$output, outf)
   if (opt$verbose) message(paste("saving results to", outf_location))
   save(list = outf, file = outf_location)
+  if (opt$mutant){
+    moutf <- gsub(".DATA", ".locustable.mutant", basename(f))
+    assign(x = moutf, mres)
+    moutf_location <- file.path(opt$output, moutf)
+    if (opt$verbose) message(paste("saving results to", moutf_location))
+    save(list = moutf, file = moutf_location)
+  }
 }
 if (opt$verbose){
   options(width = 100)